webscraping/requirements.txt
# Core backend
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29

# Data handling
pandas==2.2.2
numpy==1.26.4

# Machine Learning
scikit-learn==1.4.2
joblib==1.4.2

# Text matching
rapidfuzz==3.7.0

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1

# Utilities
python-dotenv==1.0.1
